requirements.txt
Flask
Flask-Bootstrap
Flask-Mail
Flask-Migrate
Flask-Moment
Flask-SQLAlchemy
Flask-Script
Flask-WTF
Jinja2
Mako
MarkupSafe
SQLAlchemy
WTForms
Werkzeug
alembic
blinker
itsdangerous
flask-principal
requests
python-novaclient
python-glanceclient
python-neutronclient
humanize
keystoneauth1